requirements.txt
about-time==4.2.1
absl-py==1.4.0
aiosqlite==0.19.0
alembic==1.11.1
alive-progress==3.1.3
anyio==3.7.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asn1crypto==1.5.1
astroid==2.15.5
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.3
attrs==23.1.0
audioread==3.0.0
autograd==1.5
awswrangler==3.2.1
Babel==2.12.1
backcall==0.2.0
beartype==0.14.0
beautifulsoup4==4.12.2
black==22.3.0
bleach==6.0.0
boto3==1.26.120
botocore==1.29.142
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==3.1.0
checksumdir==1.1.7
click==8.1.3
cloudpickle==2.2.1
cma==3.2.2
colorama==0.4.5
colored-traceback==0.3.0
comm==0.1.3
contourpy==1.1.0
coverage==7.2.6
cramjam==2.7.0
cryptography==40.0.2
cycler==0.11.0
dash==2.10.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2021.12.0
databricks-cli==0.17.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.6
distlib==0.3.6
distributed==2021.12.0
docker==6.1.3
docutils==0.20.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
execnet==1.9.0
executing==1.2.0
Farama-Notifications==0.0.4
fast-histogram==0.11
fastjsonschema==2.17.1
fastparquet==2023.7.0
feather-format==0.4.1
filelock==3.12.0
Flask==2.1.3
flatbuffers==23.5.26
fonttools==4.41.1
fpdf==1.7.2
fpdf2==2.5.2
fqdn==1.5.1
fsspec==2023.1.0
future==0.18.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.19.0
#google-auth-oauthlib==0.4.6
google-auth-oauthlib>=0.5,<1.1
google-pasta==0.2.0
grapheme==0.6.0
grpcio==1.54.2
gunicorn==20.1.0
gymnasium==0.28.1
h5py==3.9.0
HeapDict==1.0.1
holidays==0.25
identify==2.5.24
idna==3.4
imageio==2.8.0
importlib-metadata==6.6.0
importlib-resources==6.0.0
iniconfig==2.0.0
install==1.3.5
ipykernel==6.24.0
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
isort==5.11.5
itsdangerous==2.1.2
jax-jumpy==1.0.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
junit-xml==1.9
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter-ydoc==1.0.2
jupyter_client==8.3.0
jupyter_collaboration==1.0.1
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.23.0
kaggle==1.5.16
keras==2.13.1
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
lazy-object-proxy==1.9.0
libclang==16.0.0
librosa==0.8.0
llvmlite==0.40.1
locket==1.0.0
lockfile==0.12.2
luigi==3.3.0
lxml==4.6.5
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.1
mlflow==2.5.0
moto==4.1.10
mpl-scatter-density==0.7
mpmath==1.3.0
msgpack==1.0.5
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.2
nbformat==5.9.1
nest-asyncio==1.5.6
networkx==3.1
nodeenv==1.8.0
notebook==7.0.0
notebook_shim==0.2.3
numba==0.57.1
#numpy==1.24.4
numpy>=1.22,<=1.24.3
oauthlib==3.2.2
openpyxl==3.0.7
opt-einsum==3.3.0
overrides==7.3.1
packaging==23.1
padasip==1.2.2
#pandas==1.3.5
pandas>=1.5.0
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.0
pathspec==0.11.1
patsy==0.5.3
pexpect==4.8.0
pg8000==1.21.3
pickleshare==0.7.5
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.1
plot==0.6.5
plotly==5.14.1
pluggy==1.0.0
pooch==1.7.0
pre-commit==2.19.0
progressbar2==3.53.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.19.6
psutil==5.8.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.8.0
pylint==2.17.2
pymoo==0.6.0.1
PyMySQL==1.0.3
pynndescent==0.5.10
pyparsing==3.0.9
pytest==7.2.1
pytest-cov==4.1.0
pytest-xdist==3.3.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.7
python-slugify==8.0.1
python-utils==3.5.2
pytz==2023.3
#PyYAML==5.1.2
PyYAML>=5.3
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
querystring-parser==1.2.4
redshift-connector==2.0.910
referencing==0.30.0
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.3.1
responses==0.23.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
rpds-py==0.9.2
rsa==4.9
s3transfer==0.6.1
scikit-learn==1.1.3
scipy==1.10.1
scramp==1.4.4
seaborn==0.11.2
Send2Trash==1.8.2
shap==0.41.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
SQLAlchemy==2.0.19
sqlparse==0.4.4
stable-baselines3==2.0.0a5
stack-data==0.6.2
statsmodels==0.13.5
sympy==1.10.1
tabulate==0.8.6
tblib==1.7.0
Tcl==0.2
tenacity==8.2.2
#tensorboard==2.13.0
#tensorboard-data-server==0.7.1
#tensorflow==2.13.0
#tensorflow-estimator==2.13.0
#tensorflow-macos==2.13.0
#tensorflow-metal==1.0.1
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
toolz==0.12.0
#torch==1.11.0
#torchvision==0.15.2
torch>=1.11.0
torchvision>=0.15.2
tornado==6.2
tqdm==4.48.0
traitlets==5.9.0
tsfresh==0.17.0
typed-ast==1.5.4
typeguard==4.0.0
types-PyYAML==6.0.12.10
typing==3.7.4.3
typing_extensions==4.6.2
tzdata==2023.3
umap-learn==0.5.3
Unidecode==1.2.0
unittest-xml-reporting==3.2.0
uri-template==1.3.0
urllib3==1.26.16
virtualenv==20.23.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.0.3
widgetsnbextension==4.0.8
wrapt==1.15.0
xlrd==2.0.1
XlsxWriter==3.1.2
xlwt==1.3.0
xmltodict==0.13.0
y-py==0.6.0
ypy-websocket==0.12.1
zict==3.0.0
zipp==3.16.2
